+++ src/main/java/Duke.java
@@ -1,3 +1,9 @@
+import ui.*;
+import storage.*;
+import parser.*;
+import command.*;
+import task.*;
+import dukeexception.*;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -6,7 +12,6 @@ import java.util.Scanner;
 import java.io.File;
 
 public class Duke {
-    String filePath = "/Users/Simon/Documents/duke/src/main/java/saved.txt";
     private Storage storage;
     private TaskList tasks;
     private Ui ui;
@@ -18,6 +23,10 @@ public class Duke {
             tasks = new TaskList(storage.load());
         } catch (IOException e) {
             ui.displayLoadError();
+            File file = new File(filePath);
+            if (!file.exists()) {
+                new File(filePath);
+            }
             tasks = new TaskList();
         }
     }
@@ -44,43 +53,6 @@ public class Duke {
 
     static ArrayList<Task> list = new ArrayList<>();
 
-    private static void printFileContents(String filePath, ArrayList<Task> list) throws FileNotFoundException {
-        File f = new File(filePath); // new file for given file path
-        Scanner s = new Scanner(f);
-        while (s.hasNext()) {
-            String nextLine = s.nextLine();
-            String[] split = nextLine.split("|");
-            String indicator = split[0];
-            switch (indicator) {
-                case "D":
-                    Deadline deadline = new Deadline(split[2], split[3]);
-                    if (split[1].equals("Y")) {
-                        deadline.setCheck();
-                    }
-                    list.add(deadline);
-                    break;
-
-                case "E":
-                    Event event = new Event(split[2], split[3]);
-                    if (split[1].equals("Y")) {
-                        event.setCheck();
-                    }
-                    list.add(event);
-                    break;
-
-                case "T":
-                    Todo todo = new Todo(split[2]);
-                    if (split[1].equals("Y")) {
-                        todo.setCheck();
-                    }
-                    list.add(todo);
-                    break;
-
-                default:
-                    break;
-            }
-        }
-    }
 
     public static void main(String[] args) {
         new Duke("/Users/Simon/Documents/duke/src/main/java/duke.txt").run();
