+++ src/main/java/command/AddCommand.java
@@ -1,3 +1,7 @@
+package command;
+import task.*;
+import ui.*;
+import storage.*;
 import java.io.IOException;
 
 public class AddCommand extends Command {
