+++ src/main/java/command/DeleteCommand.java
@@ -1,16 +1,21 @@
+package command;
+import task.*;
+import ui.*;
+import storage.*;
 import java.io.IOException;
 
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
     protected int id;
 
-    public DoneCommand(int id) {
+    public DeleteCommand(int id) {
         this.id = id;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws IOException {
-        taskList.get(id - 1).setCheck();
-        ui.displayDoneTask(taskList.get(id - 1));
+        Task task = taskList.get(id - 1);
+        taskList.delete(id - 1);
+        ui.displayDeletedTask(task, taskList);
         storage.save(taskList);
     }
