+++ src/main/java/command/DoneCommand.java
@@ -1,17 +1,20 @@
+package command;
+import task.*;
+import ui.*;
+import storage.*;
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class DoneCommand extends Command {
     protected int id;
 
-    public DeleteCommand(int id) {
+    public DoneCommand(int id) {
         this.id = id;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws IOException {
-        Task task = taskList.get(id - 1);
-        taskList.delete(id - 1);
-        ui.displayDeletedTask(task, taskList);
+        taskList.get(id - 1).setCheck();
+        ui.displayDoneTask(taskList.get(id - 1));
         storage.save(taskList);
     }
