+++ src/main/java/duke.txt
@@ -1,4 +1,2 @@
-T/N/borrow book
-
-E/N/IHG /2020-02-02
+D/N/return book /2020-01-01
