+++ src/main/java/dukeexception/DukeException.java
@@ -1,3 +1,6 @@
+package dukeexception;
+
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
