+++ src/main/java/parser/Parser.java
@@ -1,3 +1,8 @@
+package parser;
+import task.*;
+import command.*;
+import dukeexception.*;
+
 public class Parser {
     public static Command parseCommand(String userInput) throws DukeException{
         String[] split = userInput.split(" ", 2);
