+++ src/main/java/task/Todo.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Todo extends Task {
 
     public Todo(String command) {
@@ -6,7 +8,7 @@ public class Todo extends Task {
 
     @Override
     public String toStringTxt() {
-        return "T/" + super.getIcon() + "/" + command + "\n";
+        return "T/" + super.getIcon() + "/" + command;
     }
 
     @Override
